토크나이징 완료(테스트 필요)

+++ 09_minishell/libft/ft_substr.c
@@ -6,17 +6,17 @@
 /*   By: yoson <yoson@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/08 16:21:01 by yoson             #+#    #+#             */
-/*   Updated: 2022/08/30 15:09:56 by yoson            ###   ########.fr       */
+/*   Updated: 2022/08/31 22:44:33 by yoson            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include "libft.h"
 
-char	*ft_substr(char const *src, unsigned int start, size_t len)
+char	*ft_substr(char const *src, int start, int len)
 {
 	char	*str;
-	size_t	i;
+	int		i;
 
 	if (ft_strlen(src) <= start)
 		return (ft_strdup(""));
